refactor(step-4): move database.ts to src/infra/database.ts
First file relocated into the infra layer. No logic change — just:
- the file now sits at src/infra/database.ts
- its own imports go up one level (../config.js, ../logging.js)
- the two operations that read from it (notes.ts, tags.ts) now import
  from ./infra/database.js while they still sit at src/ root
  (their own relocation into operations/ happens in steps 6 and 7).
