+++ src/infra/database.ts
@@ -3,8 +3,8 @@ import { homedir } from 'node:os';
 import { join } from 'node:path';
 import { existsSync } from 'node:fs';
 
-import { BEAR_DATABASE_PATH, ERROR_MESSAGES } from './config.js';
-import { logAndThrow, logger } from './logging.js';
+import { BEAR_DATABASE_PATH, ERROR_MESSAGES } from '../config.js';
+import { logAndThrow, logger } from '../logging.js';
 
 function getBearDatabasePath(): string {
   // Environment override for testing
